+++ src/pages/performance-review/PerformanceReviewPage.tsx
@@ -15,7 +15,7 @@ import {
 import { OpenAIService } from "../../business/open-ai.service";
 import { AutoTextArea } from "../../components/common/AutoTextArea";
 import { useResultState } from "../../state/result-state";
-import { ResultsError } from "../../components/results/ResultsComponent";
+import { ResultsError } from "../../components/results/ResultsError";
 import { ResultsBreadcrumbs } from "../../components/common/Breadcrumbs";
 import { useInputDetailsState } from "../../state/input-details.state";
 import { InputDetailsComponent } from "../../components/results/InputDetailsComponent";
