+++ src/pages/resumes/ReferralPage.tsx
@@ -7,7 +7,7 @@ import { ReferralLetterBreadcrumbs } from "../../components/common/Breadcrumbs";
 import { Footer } from "../../components/common/Footer";
 import { NavbarMin } from "../../components/common/NavbarMin";
 import { InputDetailsComponent } from "../../components/results/InputDetailsComponent";
-import { ResultsError } from "../../components/results/ResultsComponent";
+import { ResultsError } from "../../components/results/ResultsError";
 import {
   GenerateResultsButton,
   CopyResultsButton,
